benchmarks: Rearrange fixtures
Define benchmark fixture in benchmarks/conftest.py so that all files get
it and move cache disabling decorator to tests.common.
