+++ tests/benchmarks/test_overheads.py
@@ -1,26 +1,6 @@
 from firedrake import *
+from tests.common import disable_cache_lazy
 import pytest
-from decorator import decorator
-
-try:
-    import pytest_benchmark     # noqa: Checking for availability of plugin
-except ImportError:
-    @pytest.fixture(scope='module')
-    def benchmark():
-        return lambda c: pytest.skip("pytest-benchmark plugin not installed")
-
-
-@decorator
-def disable_cache_lazy(func, *args, **kwargs):
-    val = parameters["assembly_cache"]["enabled"]
-    parameters["assembly_cache"]["enabled"] = False
-    lazy_val = parameters["pyop2_options"]["lazy_evaluation"]
-    parameters["pyop2_options"]["lazy_evaluation"] = False
-    try:
-        func(*args, **kwargs)
-    finally:
-        parameters["assembly_cache"]["enabled"] = val
-        parameters["pyop2_options"]["lazy_evaluation"] = lazy_val
 
 
 @disable_cache_lazy
